feat: week calendar for logging hours via drag-and-drop time blocks
Adds a 5-day (Mon-Fri) hour-grid where assigned tickets sit in a palette
and users log work by drawing/dragging time blocks. Blocks snap to 15min,
move/resize/delete, support many blocks per ticket, and persist exactly as
arranged via real start/end timestamps.

Backend: TimeEntry gains nullable start_time/end_time (UTC) and hours widens
to Numeric; WorkItem hours columns widen to Numeric so fractional sums stay
correct. New /api/time-blocks router (GET week / POST / PATCH / DELETE) with
assignee-only + done-frozen authz and self-healing SUM rollups. Removed
int() truncation in capacity/pulse; LogHoursRequest + work-item response
hours are now float. Idempotent migration mirrors migrate_widen_*.

Frontend: new WeekCalendar page (pure tested lib for geometry + overlap
layout, pointer-driven drag hook, optimistic CRUD mutations, props-down
views), route /week + nav. parseInt->parseFloat on hours inputs. MSW
handlers + unit/hook tests.

+++ app/src/components/WorkItemPanel/WorkItemPanel.tsx
@@ -128,7 +128,7 @@ const WorkItemPanel = (props: WorkItemPanelProps) => {
   };
 
   const handleLogHours = () => {
-    const hours = parseInt(logHoursRef.current?.value || '0');
+    const hours = parseFloat(logHoursRef.current?.value || '0');
     if (hours <= 0) return;
     if (props.variant === 'full') {
       props.onLogHours(item, hours);
